Changes to the domain model for better structure and readability. Regular changes.
Including the removing of entities page and section which are now part of item as an option of itemType. Refactoring of Item.java, ItemType.java, ItemTypeConverter.java, ItemTypeConverter.java and ItemRepositoryTest.java, to accommodate for the storage capability of pages and sections. Adaptation of all migration scripts according to the changes in the model. Deletion of now redundant repositories, services and controllers.

Changes to the enumerators, whereas values are now written out instead of a single char representing the beginning letter of the value for better readability and semantics overall. Refactoring of AccountType.java, AccountTypeConverter.java, AccountTypeConverterTest.java ActionType.java, ActionTypeConverter.java, ActionTypeConverterTest.java, ItemType.java, ItemTypeConverter.java and ItemTypeConverterTest.java.

Other regular and structural changes.