Implementation of Server Side Rendering (SSR). Regular changes.
Implementation of ItemRouter.java which handles the routes and is responsible for loading data through ItemService.java and further ItemRepository.java.

Implementation of thymeleaf page index.html and 404.html (for non-existing routes) which is rendering the data using complex algorithmic logic and furthermore displaying them as markup on the page.

Refactoring of TestFixtures.java to accommodate for the needs of page related tests.

Regular changes.