Further Implementation of Server Side Rendering (SSR). Regular changes.
Implementation of Account.java which handles the routes and is responsible for loading data through AccountService.java and further AccountRepository.java, AccountNoContentException.java used when finding an account.

Implementation of testing for ItemRouter.java and ItemService.java.

Implementation of thymeleaf accounts pages management.html which is displaying a list of all accounts for accounts with administrational privileges to manage and 404.html (for non-existing routes).

Regular changes.