Complete basic Implementation of Server Side Rendering (SSR) and small improvements to api. Regular changes.
Refactoring of AccountForm.java and AccountCommand.java to AccountRequest.java and AccountResult.java for simplicity and avoidance of duplicate but not redundant functionalities.

Merging of account and password exceptions into AccountException.java and PasswordPolicyViolationException.java, resulting in the deletion of UserAlreadyExistsException.java. New exceptions for business case oriented handling. Refactoring to make all Exceptions base on BaseException.java.

Simplifying of delete operation for account management (management.html), proper implementation of error handling for account registration (register.html).

Proper implementation of post mapping for register.html as well as for put mapping and error handling for the service.
Separation of update and save operations in service.

Regular changes.